requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
httpx==0.28.1
websockets==14.1
networkx==3.4.2
langgraph==0.2.60
langchain==0.3.13
langchain-openai==0.2.14
autogen-agentchat==0.4.1
autogen-ext[openai]==0.4.1
python-dotenv==1.0.1
